backend/requirements.txt
fastapi==0.128.2
uvicorn[standard]==0.40.0
pydantic[email]==2.12.5
pydantic-settings==2.12.0
python-multipart==0.0.22
python-dotenv==1.2.1
sqlalchemy[asyncio]==2.0.46
asyncpg==0.31.0
alembic==1.18.3
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
psycopg2-binary
